requirements.txt
# Core ML libraries
numpy>=1.24.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# Image processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Web interface
gradio>=4.0.0

# Optional: Jupyter for notebooks
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Optional: Data visualization (for notebooks)
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
